5.6更新内容
【优化调整】
1.调整本地开发环境下前端代理配置，新增模块无需额外增加路径转发配置
2.解决jar包版本冲突
3.更新架构图和项目说明，新增elasticsearch组件集成相关内容
4.优化用户请求环节处理逻辑，避免每次请求发起多次数据库查询填充用户常用属性（组织机构名称、部门标识、公司标识、组织机构全路径等），变更为登录成功时处理一次，写入redis缓存，读取时通过redis缓存来加载
5.移除迁移过程中因实现变更不再使用的组件类别（PortletCategory）相关源码
6.重构实体基类，增加标识模型和映射模型，拆分公共属性，建立继承层次
7.调整实体配置实体类和视图模型类模板代码及代码生成处理，使其能处理父类
8.平台所有多对多关系映射表实体模型调整为映射模型，去除库表逻辑删除位，逻辑删除变更为物理删除，避免每次删除重建产生大量逻辑删除数据，影响性能
9.移除平台二次封装实现的DateUtils工具类，直接使用apache commons中的日期工具类

+++ platform-common/src/main/java/tech/abc/platform/common/utils/CacheUtil.java
@@ -46,7 +46,7 @@ public class CacheUtil {
      * @param timeout  时间
      * @param timeUnit 时间单位
      */
-    public <T> void set(String key, T value, Integer timeout, TimeUnit timeUnit) {
+    public <T> void set(String key, T value, long timeout, TimeUnit timeUnit) {
         redisTemplate.opsForValue().set(key, value, timeout, timeUnit);
     }
